requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
tqdm>=4.66.0